Algo in its own file
