+++ py/ai/h2o/sparkling/ml/algos/__init__.py
@@ -14,4 +14,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-from algos import H2OKMeans
+from kmeans import H2OKMeans
